+++ packages/playground/src/pages/about/about.vue
@@ -19,6 +19,7 @@
 				<text class="tag tag-blue">路由元信息</text>
 				<text class="tag tag-green">RouterLink 组件</text>
 				<text class="tag tag-orange">uni API 拦截</text>
+				<text class="tag tag-purple">params 参数传递</text>
 			</view>
 		</view>